verifiable_langchain_agent/pyproject.toml
[tool.poetry]
name = "verifiable-langchain-agent"
version = "0.1.1"
description = "A project to verify agent connections using TLS Notary"
authors = ["Artem <elus.aegis@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/verifiable-langchain-agent"
repository = "https://github.com/yourusername/verifiable-langchain-agent"
documentation = "https://github.com/yourusername/verifiable-langchain-agent#readme"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-core = "^0.3.18"
httpx-sse = "^0.4.0"
httpx = "^0.27.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
langchain-community = "^0.3.7"
langgraph = "^0.2.48"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
pythonpath = "src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"